make output directory structure as intended by the tooling
This means that the output when creating a project from this template will by default explode in your current directory,
but GUI apps like Visual Studio expect this to happen or the created template will have unnecessary nesting.

So, instead instruct creating a new project using the --output argument instead of --name as --output doesn't explode.